requirements.txt
category_encoders==2.5.1.post0
cflib==0.1.18.1
numpy==1.22.3
pandas<1.5.1
Pillow==9.3.0
scikit_learn==1.2.0
streamlit==1.14.0
